Move list package in `pkg` as it doesn't depend on reconcilier…
… code :)
