+++ pkg/list/diff.go
