+++ pkg/reconciler/v1alpha1/pipeline/resources/dag.go
@@ -21,7 +21,7 @@ import (
 	"strings"
 
 	"github.com/knative/build-pipeline/pkg/apis/pipeline/v1alpha1"
-	"github.com/knative/build-pipeline/pkg/reconciler/v1alpha1/taskrun/list"
+	"github.com/knative/build-pipeline/pkg/list"
 )
 
 // Node represents a Task in a pipeline.
