+++ pkg/reconciler/v1alpha1/pipeline/resources/dag_test.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/knative/build-pipeline/pkg/apis/pipeline/v1alpha1"
-	"github.com/knative/build-pipeline/pkg/reconciler/v1alpha1/taskrun/list"
+	"github.com/knative/build-pipeline/pkg/list"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
